requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_tflow_select=2.2.0=eigen
abseil-cpp=20211102.0=he9d5cce_0
absl-py=1.3.0=py310hecd8cb5_0
aiohttp=3.8.3=py310h6c40b1e_0
aiosignal=1.2.0=pyhd3eb1b0_0
anyio=3.6.2=pyhd8ed1ab_0
appnope=0.1.2=py310hecd8cb5_1001
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py310h90acd4f_3
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=py_0
async-timeout=4.0.2=py310hecd8cb5_0
attrs=22.2.0=pyh71513ae_0
babel=2.11.0=pyhd8ed1ab_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.2=pyha770c72_0
blas=1.0=mkl
bleach=6.0.0=pyhd8ed1ab_0
blinker=1.4=py310hecd8cb5_0
bottleneck=1.3.5=py310h4e76f89_0
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
brotlipy=0.7.0=py310h90acd4f_1005
bzip2=1.0.8=h1de35cc_0
c-ares=1.18.1=hca72f7f_0
ca-certificates=2023.01.10=hecd8cb5_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.12.7=py310hecd8cb5_0
cffi=1.15.1=py310ha78151a_3
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.0.4=py310hecd8cb5_0
comm=0.1.2=pyhd8ed1ab_0
contourpy=1.0.5=py310haf03e11_0
cryptography=39.0.0=py310h23bb4e6_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.6=py310h7a76584_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.3=pyhd3eb1b0_0
fftw=3.3.9=h9ed2024_1
flatbuffers=2.0.0=h23ab428_0
flit-core=3.8.0=pyhd8ed1ab_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=hd8bbffd_0
frozenlist=1.3.3=py310h6c40b1e_0
gast=0.4.0=pyhd3eb1b0_0
giflib=5.2.1=h6c40b1e_1
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpc-cpp=1.46.1=h067a048_0
grpcio=1.42.0=py310ha29bfda_0
h5py=3.7.0=py310h6c517f8_0
hdf5=1.10.6=h10fe05b_1
icu=58.2=h0a44026_3
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.0.0=pyha770c72_0
importlib_metadata=6.0.0=hd8ed1ab_0
importlib_resources=5.10.2=pyhd8ed1ab_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.21.0=pyh736e0ef_0
ipython=8.8.0=py310hecd8cb5_0
ipython_genutils=0.2.0=py_1
jedi=0.18.1=py310hecd8cb5_1
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=hca72f7f_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=4.17.3=pyhd8ed1ab_0
jupyter_client=8.0.2=pyhd8ed1ab_0
jupyter_core=5.2.0=py310h2ec42d9_0
jupyter_events=0.6.3=pyhd8ed1ab_0
jupyter_server=2.2.0=pyhd8ed1ab_0
jupyter_server_terminals=0.4.4=pyhd8ed1ab_1
jupyterlab=3.5.3=pyhd8ed1ab_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_server=2.19.0=pyhd8ed1ab_0
keras=2.10.0=py310hecd8cb5_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.4=py310hcec6c5f_0
krb5=1.19.4=hdba6334_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libcurl=7.87.0=ha585b31_0
libcxx=14.0.6=hccf4f1f_0
libdeflate=1.8=h9ed2024_5
libedit=3.1.20221030=h6c40b1e_0
libev=4.33=h9ed2024_1
libffi=3.4.2=hecd8cb5_6
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libnghttp2=1.46.0=ha29bfda_0
libpng=1.6.37=ha441bb4_0
libprotobuf=3.20.1=h8346a28_0
libsodium=1.0.18=hbcb3906_1
libssh2=1.10.0=h0a4fc7d_0
libtiff=4.5.0=hcec6c5f_1
libwebp=1.2.4=h56c3ce4_0
libwebp-base=1.2.4=hca72f7f_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.4=hcec6c5f_0
markdown=3.4.1=py310hecd8cb5_0
markupsafe=2.1.2=py310h90acd4f_0
matplotlib=3.6.2=py310hecd8cb5_0
matplotlib-base=3.6.2=py310h220de94_0
matplotlib-inline=0.1.6=py310hecd8cb5_0
mistune=2.0.4=pyhd8ed1ab_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py310hca72f7f_0
mkl_fft=1.3.1=py310hf879493_0
mkl_random=1.2.2=py310hc081a56_0
multidict=6.0.2=py310hca72f7f_0
munkres=1.1.4=py_0
nbclassic=0.4.8=pyhd8ed1ab_0
nbclient=0.7.2=pyhd8ed1ab_0
nbconvert=7.2.9=pyhd8ed1ab_0
nbconvert-core=7.2.9=pyhd8ed1ab_0
nbconvert-pandoc=7.2.9=pyhd8ed1ab_0
nbformat=5.7.3=pyhd8ed1ab_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=2.8.4=py310hecd8cb5_0
notebook=6.5.2=pyha770c72_1
notebook-shim=0.2.2=pyhd8ed1ab_0
numexpr=2.8.4=py310h9638375_0
numpy=1.23.5=py310h9638375_0
numpy-base=1.23.5=py310ha98c3c9_0
oauthlib=3.2.1=py310hecd8cb5_0
openssl=1.1.1s=hca72f7f_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.2=py310h3ea8b11_0
pandoc=2.19.2=h694c41f_1
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.3.0=py310h81888ad_1
pip=22.3.1=py310hecd8cb5_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
platformdirs=2.6.2=pyhd8ed1ab_0
prometheus_client=0.16.0=pyhd8ed1ab_0
prompt-toolkit=3.0.36=py310hecd8cb5_0
protobuf=3.20.1=py310he9d5cce_0
psutil=5.9.4=py310h90acd4f_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.4.0=py310hecd8cb5_0
pyopenssl=23.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=py310hecd8cb5_0
pyrsistent=0.19.3=py310h90acd4f_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.9=h218abb5_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.16.2=pyhd8ed1ab_0
python-flatbuffers=2.0=pyhd3eb1b0_0
python-json-logger=2.0.4=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytz=2022.7.1=pyhd8ed1ab_0
pyyaml=6.0=py310h90acd4f_5
pyzmq=25.0.0=py310hf615a82_0
re2=2022.04.01=he9d5cce_0
readline=8.2=hca72f7f_0
requests=2.28.2=pyhd8ed1ab_0
requests-oauthlib=1.3.0=py_0
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.0.2=py310hc081a56_0
scipy=1.9.3=py310h09290a1_0
seaborn=0.12.2=py310hecd8cb5_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=65.6.3=py310hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=he9d5cce_0
sniffio=1.3.0=pyhd8ed1ab_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
sqlite=3.40.1=h880c91c_0
stack_data=0.2.0=pyhd3eb1b0_0
tensorboard=2.10.0=py310hecd8cb5_0
tensorboard-data-server=0.6.1=py310h7242b5c_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.10.0=eigen_py310haa97f9e_0
tensorflow-base=2.10.0=eigen_py310h61e1807_0
tensorflow-estimator=2.10.0=py310hecd8cb5_0
termcolor=2.1.0=py310hecd8cb5_0
terminado=0.17.1=pyhd1c38e8_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.12=h5d9f67b_0
tomli=2.0.1=pyhd8ed1ab_0
tornado=6.2=py310h90acd4f_1
tqdm=4.64.1=py310hecd8cb5_0
traitlets=5.7.1=py310hecd8cb5_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.14=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py_1
websocket-client=1.5.0=pyhd8ed1ab_0
werkzeug=2.2.2=py310hecd8cb5_0
wheel=0.35.1=pyhd3eb1b0_0
wrapt=1.14.1=py310hca72f7f_0
xz=5.2.10=h6c40b1e_1
yaml=0.2.5=h0d85af4_2
yarl=1.8.1=py310hca72f7f_0
zeromq=4.3.4=he49afe7_1
zipp=3.12.0=pyhd8ed1ab_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.2=hcb37349_0
